Clean up PAR client
- Clarify what is needed in the events
- Cache discovery
- Log warning that had been commented out
